Stopwatch Practice Project
This project was to practice my usage of setInterval and useEffect.

The goal was to create a stop watch that counted up by 1 every second when the start button clicked. Then, once the start button was clicked, stop button would appear. The stop button then would pause the count and a reset buttomn would reset the state.