Edit Task functionality
Created functionality that allows the user to edit a specific task and update that task to a new value.

I also included some error handling that will not let a user update the task unless the title and description has at least one character.

Created some comments to explain code and help make it easily findable.